vim: wrap up the initial language client stuff
Fix the unicode error in the when the language server is linting the
code and returning the status. Also remove the old completion env
variable and have it enabled by default. This means I can get rid of the
old vim-rooter plugin because it is now included in the language server