vim: more advanced vim-plug install
The vim-plug install will now try to use both `curl` and `wget`. If it
can't find either executable, it will fail and exit out of vim.
Basically, you are not allowed to use vim without having one of these
installed at least. This might be inconvienent at times but it should
make it less of a hassle when setting it up for the first time.