zsh: Totally redo zsh and remove so many old things
Basically remove all of the scripts and startup files because they were
totally unneeded. Most of the ZSH config is in the zshrc now, with the
exception of a few functions.

I want to try to minimize the amount of functions in my main zshrc
because it looks junky and most of the time is only used once. This is
essentially a util, hence the new utils directory.

The prompt is totally redone too because I started noticing performance
issues on bad computers (raspberry pis').